1.0.0/requirements.txt
# Core Framework
fastapi>=0.104.0,<0.112.0
uvicorn[standard]>=0.24.0,<0.31.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.1.0

# Database
sqlalchemy>=2.0.23,<2.1.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
asyncpg>=0.29.0,<0.30.0
aiosqlite>=0.19.0,<1.0.0

# Search and Indexing
elasticsearch>=8.11.0,<9.0.0
elasticsearch-dsl>=8.11.0,<9.0.0

# Data Processing
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Text Processing
spacy>=3.7.0,<4.0.0
nltk>=3.8.0,<4.0.0
textblob>=0.17.0,<1.0.0

# Machine Learning
scikit-learn>=1.3.0,<2.0.0
torch>=2.1.0,<3.0.0
torchvision>=0.16.0,<1.0.0
transformers>=4.35.0,<5.0.0

# Web Scraping
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
aiohttp>=3.9.0,<4.0.0

# File Processing
PyPDF2>=3.0.0,<4.0.0
python-docx>=1.1.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# Visualization
matplotlib>=3.8.0,<4.0.0
plotly>=5.17.0,<6.0.0
seaborn>=0.13.0,<1.0.0

# API and Web
httpx>=0.25.0,<1.0.0
aiofiles>=23.2.0,<24.0.0
jinja2>=3.1.0,<4.0.0

# Configuration and Environment
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
toml>=0.10.0,<1.0.0

# Logging and Monitoring
structlog>=23.2.0,<24.0.0
prometheus-client>=0.19.0,<1.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0

# Security
passlib[bcrypt]>=1.7.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
cryptography>=42.0.0,<46.0.0

# Caching
redis>=5.0.0,<6.0.0
aioredis>=2.0.0,<3.0.0

# Task Queue
celery>=5.3.0,<6.0.0

# Development Tools
black>=23.11.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<8.0.0
mypy>=1.7.0,<2.0.0

# Date and Time
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3,<2025.0

# Utilities
tqdm>=4.66.0,<5.0.0
rich>=13.7.0,<14.0.0
click>=8.1.0,<9.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.4.0,<10.0.0

# XML processing
lxml>=4.9.0,<6.0.0 